fix: fix according code review
