+++ src/app/book-page/books.module.ts
@@ -1,9 +1,5 @@
 import {NgModule} from '@angular/core';
 import {SharedModule} from '../shared/shared.module';
-
-import {BookListComponent} from './components/book-list';
-import {BookItemComponent} from './components/book-item';
-import { BookComponent } from './book.component';
 import {BooksRoutingModule} from './books-routing.module';
 
 @NgModule({
@@ -13,12 +9,6 @@ import {BooksRoutingModule} from './books-routing.module';
   ],
   declarations: [
     BooksRoutingModule.components,
-    BookListComponent,
-    BookItemComponent,
-    BookComponent
   ],
-  exports: [
-    BookListComponent
-  ]
 })
 export class BooksModule {}
