+++ src/app/book-page/components/book-item/book-item.component.ts
@@ -1,4 +1,5 @@
 import {Component, Input, OnInit, EventEmitter, Output, OnDestroy} from '@angular/core';
+
 import {Book} from '../../models/book-model';
 
 @Component({
