+++ src/app/book-page/services/book.service.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@angular/core';
-import bookStorage from '../models/book-store';
 import {Book} from '../models/book-model';
 import {Observable, of} from 'rxjs';
+
+import bookStorage from '../models/book-store';
 
 @Injectable({
   providedIn: 'root'
