+++ src/app/cart-page/cart.module.ts
@@ -1,8 +1,5 @@
 import {NgModule} from '@angular/core';
 import {SharedModule} from '../shared/shared.module';
-import {CartListComponent} from './components/cart-list';
-import {CartItemComponent} from './components/cart-item';
-import { CartComponent } from './cart.component';
 import {CartRoutingModule} from './cart-routing.module';
 
 @NgModule({
@@ -12,9 +9,6 @@ import {CartRoutingModule} from './cart-routing.module';
   ],
   declarations: [
     CartRoutingModule.components,
-    CartListComponent,
-    CartItemComponent,
-    CartComponent,
   ],
 })
 export class CartModule {}
