+++ src/app/cart-page/index.ts
@@ -1,3 +1,4 @@
 export * from './models';
 export * from './services';
 export * from './cart.component';
+export * from './components';
