+++ src/app/core/constants.ts
@@ -11,3 +11,14 @@ export const ConstantsService: AppConstVersion = {
   App: 'TaskManager',
   Ver: '1.0',
 };
+
+export enum AppPath {
+  Empty = '',
+  Books = 'books',
+  ProductsList = 'products-list',
+  Product = 'product',
+  Cart = 'cart',
+  Orders = 'orders',
+  Edit = 'edit',
+  Admin = 'admin',
+}
