+++ src/app/core/services/constants.service.ts
@@ -1,4 +1,4 @@
 import {InjectionToken} from '@angular/core';
-import {AppConstVersion} from '../../models/app-const-version';
+import {AppConstVersion} from '../../models';
 
 export const CONSTANTS = new InjectionToken<AppConstVersion>('Constants');
