+++ src/app/layout/header/header.component.html
@@ -6,12 +6,11 @@
 
   <ul>
     <li routerLinkActive="active">
-      <a [routerLink]="['/books']">Книги</a>
+      <a [routerLink]="[books]">Книги</a>
     </li>
     <li routerLinkActive="active">
-      <a [routerLink]="['/cart']">Корзина</a>
+      <a [routerLink]="[cart]">Корзина</a>
     </li>
-
   </ul>
 </nav>
