--- src/app/shared/shared.constants.ts
@@ -1,10 +1,0 @@
-export enum AppPath {
-  Empty = '',
-  Books = 'books',
-  ProductsList = 'products-list',
-  Product = 'product',
-  Cart = 'cart',
-  Orders = 'orders',
-  Edit = 'edit',
-  Admin = 'admin',
-}
